requirements.txt
requests==2.32.4
python-dotenv==0.9.1